{{cookiecutter.project_slug}}/requirements.txt
aiojobs>=1.1.0
aioredis>=2.0.1
anyio>=3.6.2
async-timeout>=4.0.2
attrs>=22.2.0
exceptiongroup>=1.1.0
fastapi>=0.92.0
fastapi-plugins>=0.11.0
h11>=0.14.0
h2>=4.1.0
hiredis>=2.2.2
httpx>=0.23.3
hpack>=4.0.0
hypercorn>=0.14.3
hyperframe>=6.0.1
idna>=3.4
iniconfig>=2.0.0
packaging>=23.0
pluggy>=1.0.0
priority>=2.0.0
pydantic>=1.10.4
pytest>=7.2.1
python-json-logger>=2.0.5
redis>=4.5.1
sniffio>=1.3.0
starlette>=0.25.0
tenacity>=8.2.1
toml>=0.10.2
tomli>=2.0.1
typing-extensions>=4.4.0
wsproto>=1.2.0
requests-oauthlib>=1.3.1